[feat] Adding the ability to address a 'main' playlist that can be paused
Another plugin, such as Remote Falcon, will pause a playlist for another playlist. FPP does not send a signal on pausing/resuming a playlist. FPP also doesn't send any details about the "stopped" playlist. This would cause the "Playlist Idle Behavior" to be triggered on an interrupt, but not able to restore on a 'resume' because no signal is sent out. Therefore a resumed playlist would be in the idel behavior.

This feature allows you to define a 'main' playlist, which changes whether or not the idle behavior would be triggered. This now allows for resume to occur on that one 'main' playlist.

In the future, this could be a set of playlists, but my JS is not savy enough to design a UI for a combo box, and one playlist should be enough for this.